Initial cost estimate for ASGs. And refactoring.
